Extract ParsedFileManager
